Run classify to write the file, gather is for gathering of data. Run count.dict_builder before that.
